More Windows/general fixes... still work in progress

+++ Sources/Connection/TCP.cpp
@@ -53,7 +53,7 @@ namespace
 
 	public:
 		PacketHeader(size_t myLength)
-			: length(myLength)
+			: length(static_cast<uint32_t>(myLength))
 		{
 		}
